+++ doc/changelog.rst
@@ -16,6 +16,26 @@ Multi fit of stars
 
 * :py:meth:`~astrometry.fit_stars_in_frame` has been updated to fit
   multiple stars at the same time (see:
-  :py:meth:`~cutils.multi_fit_stars`). The star fit is now a lot more
+  :py:meth:`~cutils.multi_fit_stars`). The star fit is now way more
   robust.
 
+USNO-B1 based star detection
+----------------------------
+
+* :py:meth:`~astrometry.Astrometry.query_vizier` and
+  :py:meth:`~astrometry.Astrometry.register` have been moved from
+  :py:class:`orbs.process.Spectrum` so that the registration is part of
+  the Astrometry module (which makes more sense). This way it is now
+  possible to use a star catalogue like USNO-B1 to detect stars in the
+  cube. It is not a default behaviour because extended emission region
+  contains virtually no catalogued stars. This option can be useful for
+  galaxies to avoid the confision of HII regions and stars.
+
+v1.2
+****
+
+* :py:meth:`~cutils.multi_fit_stars` noise estimation
+  enhanced. The initial estimation of the shift has also been updated.
+
+* Minor bugs fix. This version is considered as a nearly stable
+  version ready for release.